[Clean] move packages
