+++ src/main/java/jp/azisaba/lgw/rankingdisplayer/ranking/RankingHideManager.java
@@ -1,4 +1,4 @@
-package jp.azisaba.lgw.rankingdisplayer;
+package jp.azisaba.lgw.rankingdisplayer.ranking;
 
 import java.util.UUID;
